+++ addons/yat/Plugin.cs
@@ -15,6 +15,11 @@ public partial class Plugin : EditorPlugin
 	public override void _EnterTree()
 	{
 		AddAutoloadSingleton(_name, "res://addons/yat/src/YAT.tscn");
+		AddControlToBottomPanel(
+			GD.Load<PackedScene>("res://addons/yat/src/scenes/editor_terminal/EditorTerminal.tscn").Instantiate<Control>(),
+			"Terminal"
+		);
+
 		GD.Print($"{_name} {_version} loaded!");
 		GD.PrintRich("Up to date information about YAT can be found at [url=https://github.com/MASSHUU12/godot-yat/tree/main]godot-yat[/url] on GitHub.");
 	}